# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created",
            "content": {
              "application/json": {
                "example": {
                  "id": "0673b8af-5e55-4ad7-a246-6120ac62c2cf",
                  "contentHtml": "<article><div><h2 style=\"color: red\">Strawberry title</h2><img style=\"width: 640px\" src=\"https://hips.hearstapps.com/clv.h-cdn.co/assets/15/22/1432664914-strawberry-facts1.jpg?crop=1xw:0.8332147937411095xh;center,center\" alt=\"strawberry\" /><p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quisquam omnis rerum cumque deserunt natus nam impedit consequuntur non id illo, dicta et, tenetur aliquam porro amet recusandae, sit odit vero!</p><hr /><h2>Subtitle</h2><p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Animi, a repellat. Corporis fugiat voluptatem quasi dolores optio dicta blanditiis modi inventore necessitatibus corrupti, ullam consequuntur nam atque sapiente quidem incidunt.</p></div></article>",
                  "createdAt": "2026-01-05T09:29:41.141Z",
                  "updatedAt": "2026-01-05T09:29:41.141Z"
                }
              }
            }
          },
          "400": {
            "description": "Required fields are missing",
            "content": {
              "application/json": {
                "example": {
                  "message": [
                    "contentHtml must be a string"
                  ],
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a post",
        "tags": [
          "Posts"
        ]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of elements per page (optional)",
            "schema": {
              "example": 10
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (optional)",
            "schema": {
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get posts",
            "content": {
              "application/json": {
                "example": {
                  "count": 1,
                  "rows": [
                    {
                      "id": "0673b8af-5e55-4ad7-a246-6120ac62c2cf",
                      "contentHtml": "<article><div><h2 style=\"color: red\">Strawberry title</h2><img style=\"width: 640px\" src=\"https://hips.hearstapps.com/clv.h-cdn.co/assets/15/22/1432664914-strawberry-facts1.jpg?crop=1xw:0.8332147937411095xh;center,center\" alt=\"strawberry\" /><p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quisquam omnis rerum cumque deserunt natus nam impedit consequuntur non id illo, dicta et, tenetur aliquam porro amet recusandae, sit odit vero!</p><hr /><h2>Subtitle</h2><p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Animi, a repellat. Corporis fugiat voluptatem quasi dolores optio dicta blanditiis modi inventore necessitatibus corrupti, ullam consequuntur nam atque sapiente quidem incidunt.</p></div></article>",
                      "createdAt": "2026-01-05T09:29:41.141Z",
                      "updatedAt": "2026-01-05T09:29:41.141Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Error message",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all posts",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get post",
            "content": {
              "application/json": {
                "example": {
                  "id": "0673b8af-5e55-4ad7-a246-6120ac62c2cf",
                  "contentHtml": "<article><div><h2 style=\"color: red\">Strawberry title</h2><img style=\"width: 640px\" src=\"https://hips.hearstapps.com/clv.h-cdn.co/assets/15/22/1432664914-strawberry-facts1.jpg?crop=1xw:0.8332147937411095xh;center,center\" alt=\"strawberry\" /><p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quisquam omnis rerum cumque deserunt natus nam impedit consequuntur non id illo, dicta et, tenetur aliquam porro amet recusandae, sit odit vero!</p><hr /><h2>Subtitle</h2><p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Animi, a repellat. Corporis fugiat voluptatem quasi dolores optio dicta blanditiis modi inventore necessitatibus corrupti, ullam consequuntur nam atque sapiente quidem incidunt.</p></div></article>",
                  "createdAt": "2026-01-05T09:29:41.141Z",
                  "updatedAt": "2026-01-05T09:29:41.141Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Error message",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one posts",
        "tags": [
          "Posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post updated",
            "content": {
              "application/json": {
                "example": {
                  "id": "0673b8af-5e55-4ad7-a246-6120ac62c2cf",
                  "contentHtml": "<article><div><h2 style=\"color: red\">Strawberry title</h2><img style=\"width: 640px\" src=\"https://hips.hearstapps.com/clv.h-cdn.co/assets/15/22/1432664914-strawberry-facts1.jpg?crop=1xw:0.8332147937411095xh;center,center\" alt=\"strawberry\" /><p>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quisquam omnis rerum cumque deserunt natus nam impedit consequuntur non id illo, dicta et, tenetur aliquam porro amet recusandae, sit odit vero!</p><hr /><h2>Subtitle</h2><p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Animi, a repellat. Corporis fugiat voluptatem quasi dolores optio dicta blanditiis modi inventore necessitatibus corrupti, ullam consequuntur nam atque sapiente quidem incidunt.</p></div></article>",
                  "createdAt": "2026-01-05T09:29:41.141Z",
                  "updatedAt": "2026-01-05T09:29:41.141Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Error message",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit post",
        "tags": [
          "Posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete file",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "message": "the post ff6a1126-5397-41e0-8f36-bec68f3b8a9a was deleted"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "message": "Error message",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete post",
        "tags": [
          "Posts"
        ]
      }
    },
...
    "/api/referrals": {
      "post": {
        "operationId": "ReferralsController_createReferral",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "unique ID of the referral link and ID of the invited user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReferralDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The entry was created successfully.",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "400": {
            "description": "Not found Error",
            "content": {
              "application/json": {
                "examples": {
                  "USER_WITH_ID_NOT_FOUND": {
                    "summary": "User with this id was not found",
                    "value": {
                      "message": "User with this id was not found",
                      "error": "Not Found",
                      "statusCode": 400
                    }
                  },
                  "NO_USER_WITH_THIS_LINK": {
                    "summary": "No user with this referral link was found.",
                    "value": {
                      "message": "No user with this referral link was found.",
                      "error": "Not Found",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Creating a record that a user registered using a friend's referral link",
        "tags": [
          "Referrals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fl studio API",
    "description": "Documentation for Fl studio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      },
      "refresh_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Refresh token: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "contentHtml": {
            "type": "string",
            "example": "\n            <div>\n                <h2>Title</h2>\n                <p>Text</p>\n                <img src=\"image_link\" alt=\"image description\"/>\n            </div>\n        "
          }
        },
        "required": [
          "contentHtml"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "contentHtml": {
            "type": "string",
            "example": "\n            <div>\n                <h2>Title</h2>\n                <p>Text</p>\n                <img src=\"image_link\" alt=\"image description\"/>\n            </div>\n        "
          }
        }
      },
...
      "CreateReferralDto": {
        "type": "object",
        "properties": {
          "referral": {
            "type": "string",
            "description": "Referral link",
            "example": "6229bea9-29e5-4532-8241-bd193b1ee2bf"
          },
          "refereeId": {
            "type": "string",
            "description": "User ID",
            "example": "a2ab20fc-b7de-4d91-9c59-c7d8a79ad811"
          }
        },
        "required": [
          "referral",
          "refereeId"
        ]
      },
...
    }
  }
}